[中]去除restDocs,集成devTools,更新数据库表,添加静态资源,前后端整合,账户登录注销功能,前端页面控制器,常量类更新,行程结算规则

+++ src/main/java/com/zhd/controller/UserController.java
@@ -1,17 +1,20 @@
 package com.zhd.controller;
 
-
+import com.baomidou.mybatisplus.mapper.EntityWrapper;
 import com.baomidou.mybatisplus.plugins.Page;
 import com.zhd.convert.UserConvert;
+import com.zhd.enums.UserTypeEnum;
 import com.zhd.pojo.JSONResponse;
 import com.zhd.pojo.User;
 import com.zhd.service.IUserService;
 import com.zhd.util.Constants;
+import org.apache.commons.lang.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-import java.io.IOException;
+
+import javax.servlet.http.HttpSession;
 
 /**
  * <p>
@@ -22,26 +25,26 @@ import java.io.IOException;
  * @since 2018-02-05
  */
 @RestController
-@RequestMapping("/user")
+@RequestMapping("/users")
 public class UserController extends BaseController{
     @Autowired
     private IUserService userService;
-
-    @GetMapping("{id}")
-    public JSONResponse get(User user){
-        try{
-            return renderSuccess(UserConvert.convertToVO(userService.selectById(user.getId())));
-        }catch (Exception e){
-            return renderError(e.getMessage());
-        }
-    }
 
     @GetMapping("list/{current}")
     public JSONResponse list(@PathVariable("current") int pageNum, Page<User> page) {
         try {
             if(pageNum <= 0) throw new IllegalArgumentException(Constants.ILLEGAL_ARGUMENTS);
-            return renderSuccess(UserConvert.convertToVOPageInfo(userService.selectPage(page)));
+            return renderSuccess(UserConvert.convertToVOPageInfo(userService.selectPage(page, new EntityWrapper<User>().setSqlSelect("id,name,type,status").ne("type", UserTypeEnum.MANAGER.getCode()))));
         } catch (Exception e) {
+            return renderError(e.getMessage());
+        }
+    }
+
+    @GetMapping("allStaffs")
+    public JSONResponse allStaffs(){
+        try{
+            return renderSuccess(userService.getAllStaff());
+        }catch (Exception e){
             return renderError(e.getMessage());
         }
     }
@@ -60,12 +63,16 @@ public class UserController extends BaseController{
     }
 
     @PutMapping
-    public JSONResponse update(@RequestBody @Validated(User.Update.class) User record, BindingResult bindingResult) {
+    public JSONResponse update(@RequestBody @Validated(User.Update.class) User record, BindingResult bindingResult, HttpSession session) {
         try {
             if (bindingResult.hasErrors()) {
                 return renderError(bindingResult.getFieldError().getDefaultMessage());
             } else {
-                return userService.updateById(record) ? renderSuccess(record) : renderError();
+                String currentUserId = String.valueOf(session.getAttribute("userid"));
+                if(StringUtils.isNotEmpty(currentUserId) && userService.isAdmin(currentUserId)){
+                    return userService.updateById(record) ? renderSuccess(record) : renderError();
+                }
+                else return renderError(Constants.TIP_NO_PERMISSION);
             }
         } catch (Exception e) {
             return renderError(e.getMessage());
@@ -73,13 +80,17 @@ public class UserController extends BaseController{
     }
 
     @DeleteMapping("{id}")
-    public JSONResponse delete(@Validated(User.Delete.class) User record, BindingResult bindingResult){
+    public JSONResponse delete(@Validated(User.Delete.class) User record, BindingResult bindingResult, HttpSession session){
         try{
             if(bindingResult.hasErrors()){
                 return renderError(bindingResult.getFieldError().getDefaultMessage());
             }
             else if( (record = userService.selectById(record.getId()) )  != null){
-                return userService.deleteById(record.getId()) ? renderSuccess(record) : renderError();
+                String currentUserId = String.valueOf(session.getAttribute("userid"));
+                if(StringUtils.isNotEmpty(currentUserId) && userService.isAdmin(currentUserId)){
+                    return userService.deleteById(record.getId()) ? renderSuccess(record) : renderError();
+                }
+                else return renderError(Constants.TIP_NO_PERMISSION);
             }
             else{
                 return renderError(Constants.TIP_EMPTY_DATA);
